Update to latest tcgc (#1137)
This includes support for int8, int16, and float primitive types.

+++ packages/typespec-go/test/cadlranch/type/model/inheritance/recursivegroup/recursive_client_test.go
@@ -25,13 +25,13 @@ func TestRecursiveClientGet(t *testing.T) {
 			{
 				Extension: []recursivegroup.Extension{
 					{
-						Level: to.Ptr[int32](2),
+						Level: to.Ptr[int8](2),
 					},
 				},
-				Level: to.Ptr[int32](1),
+				Level: to.Ptr[int8](1),
 			},
 			{
-				Level: to.Ptr[int32](1),
+				Level: to.Ptr[int8](1),
 			},
 		},
 	})
@@ -45,13 +45,13 @@ func TestRecursiveClientPut(t *testing.T) {
 			{
 				Extension: []recursivegroup.Extension{
 					{
-						Level: to.Ptr[int32](2),
+						Level: to.Ptr[int8](2),
 					},
 				},
-				Level: to.Ptr[int32](1),
+				Level: to.Ptr[int8](1),
 			},
 			{
-				Level: to.Ptr[int32](1),
+				Level: to.Ptr[int8](1),
 			},
 		},
 	}, nil)
